refactor: sidebar and sidebar folder area
