+++ app/dashboard/main-components/DashboardDragSidebar/index.tsx
@@ -5,9 +5,11 @@ import { GripVertical } from "lucide-react"
 import { cn } from "@/lib/utils"
 
 import { Nav } from "./Nav"
+import { SidebarFolders } from "./SidebarFolders"
 
 export const DashboardDragSidebar = () => {
   const [isCollapsed, setIsCollapsed] = useState(true)
+  const [isFolderOpen, setIsFolderOpen] = useState(false)
 
   return (
     <div
@@ -16,6 +18,12 @@ export const DashboardDragSidebar = () => {
         isCollapsed ? "w-16" : "w-64"
       )}
     >
+      <SidebarFolders
+        isCollapsed={isCollapsed}
+        isOpen={isFolderOpen}
+        setIsCollapsed={setIsCollapsed}
+        setIsOpen={setIsFolderOpen}
+      />
       <Nav isCollapsed={isCollapsed} links={dashboardSidebarData} />
 
       <button
