--- app/dashboard/main-components/SidebarFolders/index.ts
@@ -1,2 +1,0 @@
-export * from "./SidebarFolders"
-export * from "./AddOrEditFolder"
